Change dev setup
dev env needs https (for crypto feature, sw, etc) on all devices (with a
valid certificate). dev.vinisync.fr points to staging rpi, and nginx redirects
to my dev machine. Local nginx only needs to listen to port 81, in clear http
(we're in the local network). Letsencrypt is used to obtain cert for this
subdomain.